refactor: AVLListのリファクタリングと主要機能の刷新
詳細:
- `AVLList`クラスの内部構造を全面的に改良。
  - 新たなメソッド（例: `toArray`, `stream`, `indexOf`, `upperBound` など）を追加し、多様な操作対応を実現。
  - ノード管理部分を効率化し、`size`と`distinctSize`の更新ロジックを改善。
  - 冗長なコードを削除し、コアロジックの簡潔化を実現。
  - ストリームおよびイテレータ操作の強化。
- `BinarySearchTree`におけるBstIteratorを調整し、不要な重複コードを除去。
  - 新たに`BstIterator`クラスを定義し、イテレーション処理を最適化。
- `Example`クラスにおける型修正（`int` → `long`）を実施し、整合性を向上。

+++ Algorithms/PrimeNumber/src/Example.java
@@ -3,7 +3,7 @@ public class Example {
 	public static void main(String[] args) {
 		PrecomputedPrimes primes = new PrecomputedPrimes(100000000);
 		long sum = 0;
-		for (int p : primes) {
+		for (long p : primes) {
 			sum += p;
 			System.out.println(p);
 		}
